requirements.txt
# Python dependencies for personal blog on Raspberry Pi 4B
# Flask web framework with WSGI support
flask==3.1.3

# Gunicorn production WSGI server for Flask
gunicorn==21.2.0

# Environment variable loading from .env files
python-dotenv==1.0.1

# Flask-Bcrypt for password hashing (Phase 2)
flask-bcrypt==1.0.1

# Markdown rendering and syntax highlighting (Phase 3)
mistune==3.2.0
python-frontmatter==1.1.0
Pygments==2.18.0
bleach==6.1.0

# SQLite3 is included in Python standard library

# SQLAlchemy ORM and Flask-Migrate for database management (Phase 4)
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
sqlalchemy==2.0.35

# Chinese pinyin transliteration for slug generation
pypinyin==0.53.0

# Development and testing dependencies
pytest==8.3.5
pytest-flask==1.3.0